refactor: move each NoteConsumptionCost impl next to its note type
Answers PhilippGackstatter's comment on #3356 that impl T lives next to
where struct T is defined: the 12 standard and 6 agglayer impls move
out of the central costs modules into their note types' files, while
the trait, NoteCost, the lookups, and the pricer stay in the costs
modules. The created_notes docs ride along verbatim except MINT's
('typically creates') and the trait-doc rework, which land with the
doc-suggestions commit. The standards-side resolves test is upgraded to
pin each variant's cost to its own table constant, mirroring the
agglayer-side pin test (swaps between close constants would otherwise
hide inside the snapshot tests' 5% drift tolerance).
