+++ crates/miden-agglayer/src/config_note.rs
@@ -25,9 +25,11 @@ use miden_protocol::note::{
     NoteType,
     PartialNoteMetadata,
 };
+use miden_standards::note::costs::NoteConsumptionCost;
 use miden_standards::note::{NetworkAccountTarget, NoteExecutionHint};
 use miden_utils_sync::LazyLock;
 
+use crate::costs::CONFIG_AGG_BRIDGE_CONSUMPTION_CYCLES;
 use crate::{EthAddress, MetadataHash, note_script};
 
 // NOTE SCRIPT
@@ -174,6 +176,15 @@ impl ConfigAggBridgeNote {
         let assets = NoteAssets::new(vec![])?;
 
         Ok(Note::with_attachments(assets, metadata, recipient, attachments))
+    }
+}
+
+// NOTE CONSUMPTION COST
+// ================================================================================================
+
+impl NoteConsumptionCost for ConfigAggBridgeNote {
+    fn consumption_cycles() -> u32 {
+        CONFIG_AGG_BRIDGE_CONSUMPTION_CYCLES
     }
 }
