+++ crates/miden-agglayer/src/costs/mod.rs
@@ -13,72 +13,5 @@
 //! one (small drift from unrelated changes is tolerated - the pricing safety margin dwarfs
 //! it).
 
-use alloc::vec;
-use alloc::vec::Vec;
-
-use miden_protocol::note::NoteScriptRoot;
-use miden_standards::note::costs::NoteConsumptionCost;
-use miden_standards::note::{BurnNote, MintNote};
-
-use crate::{
-    B2AggNote,
-    ClaimNote,
-    ConfigAggBridgeNote,
-    DeregisterAggFaucetNote,
-    RemoveGerNote,
-    UpdateGerNote,
-};
-
 mod table;
 pub use table::*;
-
-// NOTE CONSUMPTION COST IMPLS
-// ================================================================================================
-
-impl NoteConsumptionCost for ClaimNote {
-    fn consumption_cycles() -> u32 {
-        CLAIM_CONSUMPTION_CYCLES
-    }
-
-    /// Consuming a CLAIM note creates the MINT note routed to the agglayer faucet (a network
-    /// account).
-    fn created_notes() -> Vec<NoteScriptRoot> {
-        vec![MintNote::script_root()]
-    }
-}
-
-impl NoteConsumptionCost for B2AggNote {
-    fn consumption_cycles() -> u32 {
-        B2AGG_CONSUMPTION_CYCLES
-    }
-
-    /// Consuming a B2AGG note creates the BURN note routed to the agglayer faucet (a network
-    /// account).
-    fn created_notes() -> Vec<NoteScriptRoot> {
-        vec![BurnNote::script_root()]
-    }
-}
-
-impl NoteConsumptionCost for ConfigAggBridgeNote {
-    fn consumption_cycles() -> u32 {
-        CONFIG_AGG_BRIDGE_CONSUMPTION_CYCLES
-    }
-}
-
-impl NoteConsumptionCost for DeregisterAggFaucetNote {
-    fn consumption_cycles() -> u32 {
-        DEREGISTER_AGG_FAUCET_CONSUMPTION_CYCLES
-    }
-}
-
-impl NoteConsumptionCost for UpdateGerNote {
-    fn consumption_cycles() -> u32 {
-        UPDATE_GER_CONSUMPTION_CYCLES
-    }
-}
-
-impl NoteConsumptionCost for RemoveGerNote {
-    fn consumption_cycles() -> u32 {
-        REMOVE_GER_CONSUMPTION_CYCLES
-    }
-}
