+++ crates/miden-agglayer/src/deregister_note.rs
@@ -25,9 +25,11 @@ use miden_protocol::note::{
     NoteType,
     PartialNoteMetadata,
 };
+use miden_standards::note::costs::NoteConsumptionCost;
 use miden_standards::note::{NetworkAccountTarget, NoteExecutionHint};
 use miden_utils_sync::LazyLock;
 
+use crate::costs::DEREGISTER_AGG_FAUCET_CONSUMPTION_CYCLES;
 use crate::note_script;
 
 // NOTE SCRIPT
@@ -117,5 +119,14 @@ impl DeregisterAggFaucetNote {
         let assets = NoteAssets::new(vec![])?;
 
         Ok(Note::with_attachments(assets, metadata, recipient, attachments))
+    }
+}
+
+// NOTE CONSUMPTION COST
+// ================================================================================================
+
+impl NoteConsumptionCost for DeregisterAggFaucetNote {
+    fn consumption_cycles() -> u32 {
+        DEREGISTER_AGG_FAUCET_CONSUMPTION_CYCLES
     }
 }
