+++ crates/miden-agglayer/src/update_ger_note.rs
@@ -7,8 +7,10 @@ use miden_protocol::account::AccountId;
 use miden_protocol::crypto::rand::FeltRng;
 use miden_protocol::errors::NoteError;
 use miden_protocol::note::{Note, NoteScript, NoteScriptRoot};
+use miden_standards::note::costs::NoteConsumptionCost;
 use miden_utils_sync::LazyLock;
 
+use crate::costs::UPDATE_GER_CONSUMPTION_CYCLES;
 use crate::ger_note::create_ger_note;
 use crate::{ExitRoot, note_script};
 
@@ -73,5 +75,14 @@ impl UpdateGerNote {
         rng: &mut R,
     ) -> Result<Note, NoteError> {
         create_ger_note(ger, sender_account_id, target_account_id, Self::script(), rng)
+    }
+}
+
+// NOTE CONSUMPTION COST
+// ================================================================================================
+
+impl NoteConsumptionCost for UpdateGerNote {
+    fn consumption_cycles() -> u32 {
+        UPDATE_GER_CONSUMPTION_CYCLES
     }
 }
