+++ crates/miden-standards/src/note/fee_sponsorship.rs
@@ -23,6 +23,7 @@ use miden_protocol::{Felt, Word};
 
 use super::decode_optional_block_height;
 use crate::StandardsLib;
+use crate::note::costs::{FEE_SPONSORSHIP_CONSUMPTION_CYCLES, NoteConsumptionCost};
 
 // NOTE SCRIPT
 // ================================================================================================
@@ -309,6 +310,15 @@ impl TryFrom<&[Felt]> for FeeSponsorshipNoteStorage {
         )?;
 
         Ok(Self::new(feature_note_id, reclaimer, reclaim_height))
+    }
+}
+
+// NOTE CONSUMPTION COST
+// ================================================================================================
+
+impl NoteConsumptionCost for FeeSponsorshipNote {
+    fn consumption_cycles() -> u32 {
+        FEE_SPONSORSHIP_CONSUMPTION_CYCLES
     }
 }
