+++ crates/miden-standards/src/note/pause_action.rs
@@ -21,6 +21,7 @@ use miden_protocol::utils::sync::LazyLock;
 use miden_protocol::{Felt, Word};
 
 use crate::StandardsLib;
+use crate::note::costs::{NoteConsumptionCost, PAUSE_ACTION_CONSUMPTION_CYCLES};
 
 // NOTE SCRIPT
 // ================================================================================================
@@ -230,6 +231,15 @@ impl From<PauseActionNote> for Note {
         );
 
         Note::with_attachments(NoteAssets::default(), metadata, recipient, note.attachments)
+    }
+}
+
+// NOTE CONSUMPTION COST
+// ================================================================================================
+
+impl NoteConsumptionCost for PauseActionNote {
+    fn consumption_cycles() -> u32 {
+        PAUSE_ACTION_CONSUMPTION_CYCLES
     }
 }
